[refact] Move email event classes to mail package for better organization
