+++ edukit-api/src/main/java/com/edukit/member/event/MemberEmailUpdateEvent.java
@@ -1,6 +1,6 @@
 package com.edukit.member.event;
 
-import com.edukit.core.common.event.TeacherVerificationEmailEvent;
+import com.edukit.core.common.event.mail.TeacherVerificationEmailEvent;
 
 public record MemberEmailUpdateEvent(
         String email,
