+++ edukit-core/src/main/java/com/edukit/core/common/event/mail/PasswordChangeEmailEvent.java
@@ -1,4 +1,4 @@
-package com.edukit.core.common.event;
+package com.edukit.core.common.event.mail;
 
 public interface PasswordChangeEmailEvent {
     String email();
